NEWSTART-1232 Unbind tabs service onStop.
There's an extra change as I added the recommended handler for when a service is disconnected, which needed some null checking when opening a web page. I don't know if anyone's tabs service has ever crashed, maybe that was the cause of some of the custom tabs related crashes before.