NEWSTART-1430 Unregister FCM more reliably.
Set the right kind of user to trigger it, and make the request ourselves instead of through a JS function in a page that might have been unloaded.

I know it's not ideal to create the object each time, esp. since it holds an HTTP pool; would like to look at better DI in the app as a whole. EventFetcher is a worse offender as it has a 5 connection pool but is created from scratch every time. That one in particular might be better as a ScheduledJob or one of the pre-API21 libraries available.